pyproject.toml
[project]
name = "google-ads-mcp-complete"
version = "2.0.0"
description = "Complete Google Ads API v21 MCP Server - All 29 tools fully implemented and working"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-ads>=24.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "aiofiles>=24.0.0",
    "cachetools>=5.0.0",
    "structlog>=24.0.0",
    "types-protobuf>=6.32.1.20260221",
    "google-ads-stubs>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.0.0"
]

[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m not integration')",
]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.5",
]
